abort with a helpful message if we won't be able to start a server
OSX gives out super long temp directories.  This isn't a problem until
we run into an odd 103-character limit on the names of unix sockets
(really?  really.).  Here we warn and give the user a way out of it.

I could also see just manually overriding the given base dir, but...
that's not my call to make.